+++ app/src/main/java/jp/ceed/android/mylapslogger/entity/ActivityInfo.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class ActivityInfo(
 
 	@PrimaryKey @ColumnInfo(name = "activity_id")
-	val activityId: Int,
+	val activityId: Long,
 
 	var description: String,
